move function bm25 -> base
