+++ KoPrivateGPT/retrieval/bm25_retrieval.py
@@ -7,7 +7,6 @@ from rank_bm25 import BM25Okapi
 from transformers import AutoTokenizer
 
 from KoPrivateGPT.DB.base import BaseDB
-from KoPrivateGPT.pipeline.selector import ModuleSelector
 from KoPrivateGPT.retrieval.base import BaseRetrieval
 from KoPrivateGPT.schema import Passage
 from KoPrivateGPT.utils.linker import RedisDBSingleton
@@ -26,11 +25,11 @@ class BM25Retrieval(BaseRetrieval):
     """
 
     def __init__(self, save_path: str, *args, **kwargs):
+        super().__init__()
         self.data = self.load_data(save_path)
         assert (len(self.data["tokens"]) == len(self.data["passage_id"]))
         self.save_path = save_path
         self.tokenizer = AutoTokenizer.from_pretrained("EleutherAI/polyglot-ko-1.3b")
-        self.redis_db = RedisDBSingleton()
 
     @staticmethod
     def load_data(save_path: str):
@@ -80,33 +79,3 @@ class BM25Retrieval(BaseRetrieval):
     def __tokenize(self, values: List[str]):
         tokenized = self.tokenizer(values)
         return tokenized.input_ids
-
-    @staticmethod
-    def split_ids(ids: List[UUID], db_origin_list: List[json]):
-        mongo_db_ids = [ids[i] for i, db_origin in enumerate(db_origin_list) if db_origin.db_type == "mongo_db"]
-        pickle_db_ids = [ids[i] for i, db_origin in enumerate(db_origin_list) if db_origin.db_type == "pickle_db"]
-        return mongo_db_ids, pickle_db_ids
-
-    def fetch_data(self, ids: List[UUID]) -> List[Passage]:
-        db_origin_list = self.redis_db.get_json(ids)
-        mongo_db_ids, pickle_db_ids = self.split_ids(ids, db_origin_list)
-        passage_list = []
-        self.fetch_mongo_data(mongo_db_ids, db_origin_list, passage_list)
-        self.fetch_pickle_data(pickle_db_ids, db_origin_list, passage_list)
-        return passage_list
-
-    @staticmethod
-    def fetch_mongo_data(mongo_db_ids: List[UUID], db_origin_list, passage_list):
-        if mongo_db_ids:
-            mongo_db = ModuleSelector("db").select("mongo_db").get(**db_origin_list[0].db_path)
-            mongo_db.load()
-            mongo_db_passage_list = mongo_db.fetch(mongo_db_ids)
-            passage_list.append(mongo_db_passage_list)
-
-    @staticmethod
-    def fetch_pickle_data(pickle_db_ids: List[UUID], db_origin_list, passage_list):
-        if pickle_db_ids:
-            pickle_db = ModuleSelector("db").select("pickle_db").get(**db_origin_list[0].db_path)
-            pickle_db.load()
-            pickle_db_passage_list = pickle_db.fetch(pickle_db_ids)
-            passage_list.append(pickle_db_passage_list)
